requirements.txt
fastapi
uvicorn
pydantic
ultralytics
opencv-python
numpy
pillow
torch
python-multipart
dependency-injector
pydantic-settings
